% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcindex-package.R
\docType{package}
\name{fcindex-package}
\alias{fcindex-package}
\title{Fast concordance-index computation}
\description{
This package provides one function \code{CIndex}, that implements a fast
and efficient algorithm for computing concordance.
}
\keyword{package}
\keyword{regression}
\keyword{survival}
